ErrorCode can be a simple int
Reviewed-on: https://gerrit.libreoffice.org/44035
Tested-by: Caolán McNamara <caolanm@redhat.com>

+++ filter/source/graphicfilter/ios2met/ios2met.cxx
@@ -327,7 +327,7 @@ class OS2METReader {
 
 private:
 
-    long ErrorCode;
+    int ErrorCode;
 
     SvStream      * pOS2MET;             // the OS2MET file to be read
     VclPtr<VirtualDevice> pVirDev;       // here the drawing methods are being called
